DLTS & better restart condition & part of potential
1) DLTS: new function that will calculate capcitance during the
transient
2) Restart: right know when we are restarting simulation from some state
we will consider fact, that we need to take two steps back in algorithm
to be sure that we do not exceed ideal solution. At time i) we are good,
i+1) we exceed i+2) algorithm will try to came back and we will have an
information that we achieved steady state but accuracy was not good
enough, ergo we need to came back to point i)
3) I added function which calculate what part of potential is on
Schottky and what part is on pn junction. It is not necessary but it
helps me to compare this solution to the idealistic one.